+++ cate/util/im/utils.py
@@ -140,14 +140,16 @@ def compute_tile_size(total_size,
 def get_chunk_size(array):
     chunk_size = None
     try:
-        # xarray DataArray
-        chunk_size = array.encoding['chunksizes']
+        # xarray DataArray with dask, returns the size of each individual tile
+        chunk_size = array.chunks
+        if chunk_size:
+            chunk_size = tuple([c[0] if isinstance(c, tuple) else  c for c in chunk_size])
     except:
         pass
     if not chunk_size:
         try:
-            # netCDF4 data array
-            chunk_size = array.chunks
+            # netcdf 4
+            chunk_size = array.encoding['chunksizes']
         except:
-            pass
+             pass
     return chunk_size