+++ CHANGES.md
@@ -5,6 +5,7 @@
 ### Issues Fixed/Resolved
 
 * Fixed [#309](https://github.com/CCI-Tools/cate-core/issues/309)
+* Ensure that our tile size matches the expected tile size: resize and fill in background value.
 
 ## Changes in version 0.9.0.dev4